kmpcmanager broken out
