+++ kmpc/managerlibrarypanel.py
@@ -21,7 +21,7 @@ import os
 from pkg_resources import resource_filename
 
 from kmpc.extra import KmpcHelpers
-import kmpc.kmpcmanager
+import kmpc.managerinterface
 
 Helpers=KmpcHelpers()
 
@@ -36,28 +36,28 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
         self.rbl.clear_selection()
         if value == 'Files':
             self.current_view = {'value': 'root','base':'/','info':{'type':'uri'}}
-            kmpc.kmpcmanager.mainmpdconnection.protocol.lsinfo(self.current_view['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.lsinfo(self.current_view['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
             self.ids.files_button.state='down'
             self.ids.albums_button.state='normal'
             self.ids.tracks_button.state='normal'
             self.ids.playlists_button.state='normal'
         elif value == 'Albums':
             self.current_view = {'value': 'All Album Artists','base':'All Album Artists','info':{'type':'rootalbums'}}
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('albumartistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('albumartistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
             self.ids.files_button.state='normal'
             self.ids.albums_button.state='down'
             self.ids.tracks_button.state='normal'
             self.ids.playlists_button.state='normal'
         elif value == 'Tracks':
             self.current_view = {'value': 'All Track Artists','base':'All Track Artists','info':{'type':'roottracks'}}
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('artistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('artistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
             self.ids.files_button.state='normal'
             self.ids.albums_button.state='normal'
             self.ids.tracks_button.state='down'
             self.ids.playlists_button.state='normal'
         elif value == 'Playlists':
             self.current_view = {'value':'All Playlists','base':'All Playlists','info':{'type':'playlist'}}
-            kmpc.kmpcmanager.mainmpdconnection.protocol.listplaylists().addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.listplaylists().addCallback(self.reload_view).addErrback(self.handle_mpd_error)
             self.ids.files_button.state='normal'
             self.ids.albums_button.state='normal'
             self.ids.tracks_button.state='normal'
@@ -69,7 +69,7 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
             rr['copy_flag']=str(result)
         else:
             rr['copy_flag']=''
-        kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_get('song',rr['base'],'rating').addCallback(partial(self.render_row2,rr,True)).addErrback(partial(self.render_row2,rr,False))
+        kmpc.managerinterface.mainmpdconnection.protocol.sticker_get('song',rr['base'],'rating').addCallback(partial(self.render_row2,rr,True)).addErrback(partial(self.render_row2,rr,False))
 
     def render_row2(self,r,has_sticker,result):
         rr=deepcopy(r)
@@ -124,7 +124,7 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
             elif 'file' in row:
                 Logger.debug("FileBrowser: file found: ["+row['file']+"]")
                 r={'value':Helpers.formatsong(row),'base':row['file'],'info':{'type':'file'}}
-                kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_get('song',row['file'],'copy_flag').addCallback(partial(self.render_row,r,True)).addErrback(partial(self.render_row,r,False))
+                kmpc.managerinterface.mainmpdconnection.protocol.sticker_get('song',row['file'],'copy_flag').addCallback(partial(self.render_row,r,True)).addErrback(partial(self.render_row,r,False))
             else:
                 if self.current_view['info']['type'] == 'rootalbums':
                     Logger.debug("Library: album artist found: ["+row+"]")
@@ -149,17 +149,17 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
         Logger.debug("Library: handle_double_click("+format(row)+")")
         self.current_view = deepcopy(row)
         if row['info']['type'] == 'uri':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.lsinfo(row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.lsinfo(row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'rootalbums':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('albumartistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('albumartistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'albumartistsort':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('album','albumartistsort',row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('album','albumartistsort',row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'album':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.find('album',row['base'],'albumartistsort',row['info']['albumartistsort']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.find('album',row['base'],'albumartistsort',row['info']['albumartistsort']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'roottracks':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('artistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('artistsort').addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'artistsort':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.list('title','artistsort',row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.list('title','artistsort',row['base']).addCallback(self.reload_view).addErrback(self.handle_mpd_error)
         elif row['info']['type'] == 'playlist':
             pass
         elif row['info']['type'] == 'file':
@@ -178,19 +178,19 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
         for rrow in result:
             if copy_flag:
                 Logger.debug("set_copy_flag_find: setting copy_flag to "+copy_flag+" for file "+rrow['file'])
-                kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_set('song',rrow['file'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.sticker_set('song',rrow['file'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
             else:
                 Logger.debug("set_copy_flag_find: clearing copy_flag for file "+rrow['file'])
-                kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_delete('song',rrow['file'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.sticker_delete('song',rrow['file'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
 
     def set_copy_flag_find_one(self,copy_flag,index,result):
         for rrow in result:
             if copy_flag:
                 Logger.debug("set_copy_flag_find_one: setting copy_flag to "+copy_flag+" for file "+rrow['file'])
-                kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_set('song',rrow['file'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.sticker_set('song',rrow['file'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
             else:
                 Logger.debug("set_copy_flag_find_one: clearing copy_flag for file "+rrow['file'])
-                kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_delete('song',rrow['file'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.sticker_delete('song',rrow['file'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
             break
 
     def set_copy_flag(self,copy_flag):
@@ -203,21 +203,21 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
                 Logger.debug("set_copy_flag: adding uri or file")
                 if copy_flag:
                     Logger.debug("set_copy_flag: setting copy_flag to "+copy_flag+" for file "+row['base'])
-                    kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_set('song',row['base'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
+                    kmpc.managerinterface.mainmpdconnection.protocol.sticker_set('song',row['base'],'copy_flag',copy_flag).addCallback(partial(self.reload_row_after_sticker,copy_flag,index)).addErrback(self.handle_mpd_error)
                 else:
                     Logger.debug("set_copy_flag: clearing copy_flag for file "+row['base'])
-                    kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_delete('song',row['base'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
+                    kmpc.managerinterface.mainmpdconnection.protocol.sticker_delete('song',row['base'],'copy_flag').addCallback(partial(self.reload_row_after_sticker,'',index)).addErrback(self.handle_mpd_error)
             elif mtype == 'albumartistsort':
-                kmpc.kmpcmanager.mainmpdconnection.protocol.find(mtype,row['base']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.find(mtype,row['base']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
             elif mtype == 'album':
-                kmpc.kmpcmanager.mainmpdconnection.protocol.find(mtype,row['base'],'albumartistsort',row['info']['albumartistsort']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.find(mtype,row['base'],'albumartistsort',row['info']['albumartistsort']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
             elif mtype == 'artistsort':
-                kmpc.kmpcmanager.mainmpdconnection.protocol.find(mtype,row['base']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.find(mtype,row['base']).addCallback(partial(self.set_copy_flag_find,copy_flag,index)).addErrback(self.handle_mpd_error)
             elif mtype == 'track':
-                kmpc.kmpcmanager.mainmpdconnection.protocol.find('artistsort',row['info']['artistsort'],'title',row['base']).addCallback(partial(self.set_copy_flag_find_one,copy_flag,index)).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.find('artistsort',row['info']['artistsort'],'title',row['base']).addCallback(partial(self.set_copy_flag_find_one,copy_flag,index)).addErrback(self.handle_mpd_error)
             elif mtype == 'playlist' or mtype == 'uri':
                 Logger.info("Library: "+mtype+" copy_flag not implemented")
-                #kmpc.kmpcmanager.mainmpdconnection.protocol.load(row['base'])
+                #kmpc.managerinterface.mainmpdconnection.protocol.load(row['base'])
             else:
                 Logger.warning("Library: "+mtype+' copy_flag not implemented')
         self.rbl.clear_selection()
@@ -225,7 +225,7 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
     def rating_set(self,base,index,instance):
         Logger.debug('Application: rating_set('+instance.rating+')')
         instance.popup.dismiss()
-        kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_set('song',base,'rating',instance.rating).addCallback(partial(self.handle_rating_set,index,instance.rating,True)).addErrback(partial(self.handle_rating_set,index,instance.rating,False))
+        kmpc.managerinterface.mainmpdconnection.protocol.sticker_set('song',base,'rating',instance.rating).addCallback(partial(self.handle_rating_set,index,instance.rating,True)).addErrback(partial(self.handle_rating_set,index,instance.rating,False))
 
     def handle_rating_set(self,index,rating,succ,result):
         if succ:
@@ -238,7 +238,7 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
     def generate_list(self,ltype):
         Logger.info('generate_list: generating with minimum stars '+self.ids.minimum_stars.text)
         self.tlist={}
-        kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_find('song','','rating').addCallback(partial(self.generate_play_list,ltype)).addErrback(self.handle_mpd_error)
+        kmpc.managerinterface.mainmpdconnection.protocol.sticker_find('song','','rating').addCallback(partial(self.generate_play_list,ltype)).addErrback(self.handle_mpd_error)
 
     def generate_play_list(self,ltype,result):
         Logger.debug("generate_play_list: "+ltype)
@@ -250,11 +250,11 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
                 self.tlist[uri]=1
         if ltype=='playlist':
             Logger.info("generate_play_list: writing to playlist ["+self.ids.minimum_stars.text+" star or more]")
-            kmpc.kmpcmanager.mainmpdconnection.protocol.playlistclear(self.ids.minimum_stars.text+" star or more").addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.playlistclear(self.ids.minimum_stars.text+" star or more").addErrback(self.handle_mpd_error)
             for k in sorted(self.tlist.keys()):
-                kmpc.kmpcmanager.mainmpdconnection.protocol.playlistadd(self.ids.minimum_stars.text+" star or more",k).addErrback(self.handle_mpd_error)
+                kmpc.managerinterface.mainmpdconnection.protocol.playlistadd(self.ids.minimum_stars.text+" star or more",k).addErrback(self.handle_mpd_error)
         elif ltype=='synclist':
-            kmpc.kmpcmanager.mainmpdconnection.protocol.sticker_find('song','','copy_flag').addCallback(self.generate_sync_list).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.sticker_find('song','','copy_flag').addCallback(self.generate_sync_list).addErrback(self.handle_mpd_error)
 
     def generate_sync_list(self,result):
         for row in result:
@@ -269,9 +269,9 @@ class ManagerLibraryTabbedPanelItem(TabbedPanelItem):
                 except KeyError:
                     pass
         Logger.info("generate_sync_list: writing to playlist ["+self.config.get('sync','syncplaylist')+"]")
-        kmpc.kmpcmanager.mainmpdconnection.protocol.playlistclear(self.config.get('sync','syncplaylist')).addErrback(self.handle_mpd_error)
+        kmpc.managerinterface.mainmpdconnection.protocol.playlistclear(self.config.get('sync','syncplaylist')).addErrback(self.handle_mpd_error)
         for k in sorted(self.tlist.keys()):
-            kmpc.kmpcmanager.mainmpdconnection.protocol.playlistadd(self.config.get('sync','syncplaylist'),k).addErrback(self.handle_mpd_error)
+            kmpc.managerinterface.mainmpdconnection.protocol.playlistadd(self.config.get('sync','syncplaylist'),k).addErrback(self.handle_mpd_error)
 
 class LibraryRecycleBoxLayout(FocusBehavior,LayoutSelectionBehavior,RecycleBoxLayout):
     ''' Adds selection and focus behaviour to the view. '''
